Get compiling on win32 again
Next up: moving the game code into dll like on linux

+++ src/core/allocator.cpp
@@ -301,7 +301,7 @@ T *ialloc(A *a)
 }
 
 template <typename T, typename A>
-T *alloc_array(A *a, i32 count)
+T *alloc_array(A *a, isize count)
 {
 	return (T*)alloc(a, sizeof(T) * count);
 }
